requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
mysqlclient==2.0.0
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
